# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi",
        "description": "Laravel OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "local server"
        },
        {
            "url": "https://example.com/",
            "description": "staging server"
        },
        {
            "url": "https://api.example.com/",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/forms/{formId}/questions-and-answers": {
            "get": {
                "tags": [
                    "Forms"
                ],
                "summary": "Obtenir preguntes i respostes d'un formulari",
                "operationId": "66c794752e43f66ca91ba300257b466a",
                "parameters": [
                    {
                        "name": "formId",
                        "in": "path",
                        "description": "ID del formulari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Preguntes i respostes obtingudes correctament"
                    },
                    "404": {
                        "description": "Formulari no trobat"
                    }
                }
            }
        },
        "/api/forms": {
            "get": {
                "tags": [
                    "Forms"
                ],
                "summary": "Llistar tots els formularis",
                "operationId": "f3b25f912126359aa20da08d3e875714",
                "responses": {
                    "200": {
                        "description": "Llista de formularis obtinguda correctament"
                    }
                }
            },
            "post": {
                "tags": [
                    "Forms"
                ],
                "summary": "Crear un nou formulari",
                "operationId": "6cd3f97447c1b326d440894389e4459e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Formulari d'exemple"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Formulari creat correctament"
                    },
                    "422": {
                        "description": "Error de validació"
                    }
                }
            }
        },
        "/api/forms/{id}": {
            "get": {
                "tags": [
                    "Forms"
                ],
                "summary": "Obtenir un formulari específic per ID",
                "operationId": "17035b81e30778da2bcbbaa959fef9b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del formulari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dades del formulari"
                    },
                    "404": {
                        "description": "Formulari no trobat"
                    }
                }
            },
            "put": {
                "tags": [
                    "Forms"
                ],
                "summary": "Actualitzar un formulari existent",
                "operationId": "f229471e4a4f98827ba74d73e236ebb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del formulari a actualitzar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Títol actualitzat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formulari actualitzat correctament"
                    },
                    "400": {
                        "description": "Error de validació"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Forms"
                ],
                "summary": "Eliminar un formulari",
                "operationId": "e9712e2f46a318b3438714b125cd518f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del formulari a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Formulari eliminat correctament"
                    },
                    "404": {
                        "description": "Formulari no trobat"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Autorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}